started adding html elements for different overview features
