+++ package.json
@@ -14,7 +14,7 @@
     "react-dom": "^18.2.0",
     "react-icons": "^5.0.1",
     "react-pro-sidebar": "^1.1.0",
-    "react-scripts": "5.0.1",
+    "react-scripts": "^5.0.1",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
